requirements.txt
selenium
pyyaml
python-dotenv
webdriver-manager
